INatsConnection rename and namespace changes
Also removed flush, observable and sharding.
