+++ src/NATS.Client.Core/Internal/NatsReadProtocolProcessor.cs
@@ -8,9 +8,8 @@ using System.Text;
 using System.Text.Json;
 using Microsoft.Extensions.Logging;
 using NATS.Client.Core.Commands;
-using NATS.Client.Core.Internal;
 
-namespace NATS.Client.Core;
+namespace NATS.Client.Core.Internal;
 
 internal sealed class NatsReadProtocolProcessor : IAsyncDisposable
 {
